do not use poetry.lock during docker build +ignore
Due to our service / service host deployment model and due to the way we
install dependencies in the docker image, poetry.lock was irrelevant. We
were always refreshing it from the .toml files.

We may want to reintroduce the usage of poetry.lock in the future, yet
we will need to find a way to make poetry.lock work with our service /
service host deployment model. And the constraints in the .toml files
will likely have to be relaxed (to get the loose constraints in .toml
and strict ones in the form of the .lock file)

Until then, poetry.lock will only be used locally, so:
- remove poetry.lock from the rococo-service-base-onbuild image
- git ignore poetry.lock