add hint on rococo_svchost.logger module resolution in PyCharm
The rococo_svchost.logger resolution issue in PyCharm is specific to
this git repository, since the name of the directory (rococo_svchost)
shadows the imported rococo_svchost module if the root of the repository
is opened as a PyCharm project.

"Mark directory as sources root" adds the directory to the PATH(or
PYTHONPATH) in the PyCharm environment, allowing PyCharm to resolve the
module properly.

In this case, this is for the resolution only, since service modules are
not supposed to be run with PyCharm directly ('poetry run svchost' is
used instead)